at-commander-module/at-commander-module-quectel/src/main/java/com/github/pmoerenhout/atcommander/module/quectel/commands/IccidResponse.java
package com.github.pmoerenhout.atcommander.module.quectel.commands;

import java.util.List;
import java.util.regex.Matcher;
import java.util.regex.Pattern;

import com.github.pmoerenhout.atcommander.AtResponse;
import com.github.pmoerenhout.atcommander.basic.commands.BaseResponse;
import com.github.pmoerenhout.atcommander.basic.commands.Response;

public class IccidResponse extends BaseResponse implements Response {

  private static final Pattern PATTERN = Pattern.compile("^\\+QCCID: (\\d*)(F?)$");

  private String iccid;

  public IccidResponse(final AtResponse s) {
    parseSolicited(s);
  }

  public void parseSolicited(final AtResponse response) {
    final List<String> informationalText = response.getInformationalText();
    if (informationalText.size() == 1) {
      final String line = informationalText.get(0);
      final Matcher m = PATTERN.matcher(line);
      if (m.find()) {
        iccid = m.group(1);
        return;
      }
      throw createParseException(line);
    }
    throw createParseException(response);
  }

  public String getIccid() {
    return iccid;
  }
}
